update move descriptions tweak tm descriptions

+++ include/item.h
@@ -77,7 +77,7 @@ u8 ItemId_GetType(u16 itemId);
 ItemUseFunc ItemId_GetFieldFunc(u16 itemId);
 u8 ItemId_GetBattleUsage(u16 itemId);   //not using now, usnig battlfunc check instead
 ItemUseFunc ItemId_GetBattleFunc(u16 itemId);
-u8 ItemId_GetSecondaryId(u16 itemId);
+u16 ItemId_GetSecondaryId(u16 itemId);
 u32 ItemId_GetFlingPower(u32 itemId);
 u16 itemid_get_market_price(u16 itemId);
 u16 itemid_get_Id(u16 itemId);
